Fix for missing 'trackComplete' event when application destroys player before it receives the 'playbackfinished' from player
The adobe analytics plugin registers the 'playbackbackfinished' callback in 'sourceloaded' callback. When application registers the event listeners before loading the source, the 'playbackfinished' callback of application is registered before the 'playbackfinished' callback registered by adobe analytics plugin. In this scenario if application destroys the adobe plugin on 'playbackfinished', then the plugin does not get chance to send 'trackComplete' leaving the adobe heartbeat session in dangling state resulting in erros.

The fix is to register 'plybackfinished' callback when the plugin is created and asking application to register events after creating the plugin object.